+++ packages/Modal/src/components/Content/Content.styles.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+import stylers from "@paprika/stylers";
+import { spacer } from "@paprika/stylers/lib/helpers";
+
+const SPACE = spacer(2);
+
+export const ContentWrapper = styled.div`
+  overflow-y: auto;
+
+  &:focus {
+    ${stylers.focusRing.subtle(true)};
+  }
+`;
+
+export const Content = styled.div`
+  padding: ${SPACE};
+`;